Initial attempt at refactoring out a root view controller from the sitemap view
